refactor(ember): Extract template compiler plugin and refactor it.
